Update Medicines model to change price type from Decimal to Integer; enhance MedicineDto validation; update order management endpoints and adjust Prisma migrations

+++ prisma/migrations/20250126194654_pharmacy_models/migration.sql
@@ -27,7 +27,7 @@ CREATE TABLE "Medicines" (
     "id" SERIAL NOT NULL,
     "medicinieName" TEXT NOT NULL,
     "Description" TEXT NOT NULL,
-    "price" DECIMAL(65,30) NOT NULL,
+    "price" INTEGER NOT NULL,
     "stock" INTEGER NOT NULL,
     "createdAt" TIMESTAMP(3) NOT NULL DEFAULT CURRENT_TIMESTAMP,
     "updatedAt" TIMESTAMP(3) NOT NULL,
